+++ .env
@@ -1,6 +1,6 @@
 #REACT_APP_BACKEND_URL=https://rehub-backend.onrender.com/api/v1
 
-REACT_APP_BACKEND_URL=https://rehub-backend.onrender.com/api/v1
+REACT_APP_BACKEND_URL=https://localhost:8080/api/v1
 REACT_APP_RECAPTCHA_KEY=6LdbMxApAAAAAABTOlpcf-EGSBBjZnVRKI5FEKNa
 REACT_APP_VERSION=$npm_package_version
 REACT_APP_NAME=$npm_package_name
